+++ packages/vrender-core/src/graphic/graphic.ts
@@ -1,6 +1,6 @@
 import type { ICustomPath2D } from './../interface/path';
 import type { Dict, IPointLike, IAABBBounds, IOBBBounds } from '@visactor/vutils';
-import { interpolateNumber, last, OBBBounds } from '@visactor/vutils';
+import { isArray, OBBBounds } from '@visactor/vutils';
 import {
   AABBBounds,
   Matrix,
@@ -40,7 +40,7 @@ import type {
 import { EventTarget, CustomEvent } from '../event';
 import { DefaultTransform } from './config';
 import { application } from '../application';
-import { Animate, DefaultStateAnimateConfig } from '../animate';
+import { Animate, DefaultStateAnimateConfig, defaultTimeline } from '../animate';
 import { interpolateColor } from '../color-string/interpolate';
 import { CustomPath2D } from '../common/custom-path2d';
 import { ResourceLoader } from '../resource-loader/loader';
@@ -823,6 +823,7 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
     }
     this.animates.set(animate.id, animate);
     animate.onRemove(() => {
+      animate.stop();
       this.animates.delete(animate.id);
     });
 
@@ -962,13 +963,14 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
     this.normalAttrs = null;
   }
 
-  removeState(stateName: string, hasAnimation?: boolean) {
-    const index = this.currentStates ? this.currentStates.indexOf(stateName) : -1;
+  removeState(stateName: string | string[], hasAnimation?: boolean) {
+    if (this.currentStates) {
+      const filter = isArray(stateName) ? (s: string) => !stateName.includes(s) : (s: string) => s !== stateName;
+      const newStates = this.currentStates.filter(filter);
 
-    if (index >= 0) {
-      const currentStates = this.currentStates.filter(state => state !== stateName);
-
-      this.useStates(currentStates, hasAnimation);
+      if (newStates.length !== this.currentStates.length) {
+        this.useStates(newStates, hasAnimation);
+      }
     }
   }
 
@@ -1214,9 +1216,12 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
       this.layer = layer;
       this.setStageToShadowRoot(stage, layer);
       if (this.animates && this.animates.size) {
+        // 设置timeline为所属的stage上的timeline，但如果timeline并不是默认的timeline，就不用覆盖
         const timeline = stage.getTimeline();
         this.animates.forEach(a => {
-          a.setTimeline(timeline);
+          if (a.timeline === defaultTimeline) {
+            a.setTimeline(timeline);
+          }
         });
       }
       this._onSetStage && this._onSetStage(this, stage, layer);
@@ -1437,6 +1442,7 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
   detachShadow() {
     if (this.shadowRoot) {
       this.addUpdateBoundTag();
+      this.shadowRoot.release(true);
       this.shadowRoot = null;
     }
   }
@@ -1565,6 +1571,7 @@ export abstract class Graphic<T extends Partial<IGraphicAttribute> = Partial<IGr
 
   release(): void {
     this.releaseStatus = 'released';
+    this.stopAnimates();
     application.graphicService.onRelease(this);
   }