+++ packages/vrender-core/src/interface/graphic/area.ts
@@ -4,11 +4,29 @@ import type { IConnectedStyle, IGraphic, IGraphicAttribute } from '../graphic';
 import type { ICurveType } from '../common';
 
 export type IAreaAttribute = {
-  segments: IAreaSegment[]; // 分段设置point和样式
+  /**
+   * 分段设置point和样式
+   */
+  segments: IAreaSegment[];
+  /**
+   * 所有的点坐标
+   */
   points: IPointLike[];
+  /**
+   * 曲线的类型，默认为linear
+   */
   curveType: ICurveType;
+  /**
+   * 线段的裁切比例/显示长度占总长度的比例
+   */
   clipRange: number;
+  /**
+   * 是否是闭合曲线
+   */
   closePath: boolean;
+  /**
+   * 曲线类型为catmullRom时，对应的张力参数
+   */
   curveTension: number;
 };