+++ packages/vrender-core/src/interface/graphic/circle.ts
@@ -1,8 +1,17 @@
 import type { IGraphicAttribute, IGraphic } from '../graphic';
 
 export type ICircleAttribute = {
+  /**
+   * 半径
+   */
   radius: number;
+  /**
+   * 起始角度，单位为弧度
+   */
   startAngle: number;
+  /**
+   * 终止角度，单位为弧度
+   */
   endAngle: number;
 };