+++ packages/vrender-core/src/interface/graphic/polygon.ts
@@ -2,9 +2,17 @@ import type { IPointLike } from '@visactor/vutils';
 import type { IGraphicAttribute, IGraphic } from '../graphic';
 
 export type IPolygonAttribute = {
-  // ? 不需要x2/y2
+  /**
+   * 多边形的顶点坐标
+   */
   points: IPointLike[];
+  /**
+   * 圆角半径
+   */
   cornerRadius?: number | number[];
+  /**
+   * 是否闭合多边形
+   */
   closePath?: boolean;
 };