+++ packages/vrender/__tests__/browser/src/pages/vchart.ts
@@ -20,91 +20,59 @@ export const page = () => {
     let i = 0;
 
     const spec = {
-      type: 'area',
-      data: {
-        values: [
-          { type: 'Nail polish', country: 'Africa', value: 4229 },
-          { type: 'Nail polish', country: 'EU', value: 4376 },
-          { type: 'Nail polish', country: 'China', value: 3054 },
-          { type: 'Nail polish', country: 'USA', value: 12814 },
-          { type: 'Eyebrow pencil', country: 'Africa', value: 3932 },
-          { type: 'Eyebrow pencil', country: 'EU', value: 3987 },
-          { type: 'Eyebrow pencil', country: 'China', value: 5067 },
-          { type: 'Eyebrow pencil', country: 'USA', value: 13012 },
-          { type: 'Rouge', country: 'Africa', value: 5221 },
-          { type: 'Rouge', country: 'EU', value: 3574 },
-          { type: 'Rouge', country: 'China', value: 7004 },
-          { type: 'Rouge', country: 'USA', value: 11624 },
-          { type: 'Lipstick', country: 'Africa', value: 9256 },
-          { type: 'Lipstick', country: 'EU', value: 4376 },
-          { type: 'Lipstick', country: 'China', value: 9054 },
-          { type: 'Lipstick', country: 'USA', value: 8814 },
-          { type: 'Eyeshadows', country: 'Africa', value: 3308 },
-          { type: 'Eyeshadows', country: 'EU', value: 4572 },
-          { type: 'Eyeshadows', country: 'China', value: 12043 },
-          { type: 'Eyeshadows', country: 'USA', value: 12998 },
-          { type: 'Eyeliner', country: 'Africa', value: 5432 },
-          { type: 'Eyeliner', country: 'EU', value: 3417 },
-          { type: 'Eyeliner', country: 'China', value: 15067 },
-          { type: 'Eyeliner', country: 'USA', value: 12321 },
-          { type: 'Foundation', country: 'Africa', value: 13701 },
-          { type: 'Foundation', country: 'EU', value: 5231 },
-          { type: 'Foundation', country: 'China', value: 10119 },
-          { type: 'Foundation', country: 'USA', value: 10342 },
-          { type: 'Lip gloss', country: 'Africa', value: 4008 },
-          { type: 'Lip gloss', country: 'EU', value: 4572 },
-          { type: 'Lip gloss', country: 'China', value: 12043 },
-          { type: 'Lip gloss', country: 'USA', value: 22998 },
-          { type: 'Mascara', country: 'Africa', value: 18712 },
-          { type: 'Mascara', country: 'EU', value: 6134 },
-          { type: 'Mascara', country: 'China', value: 10419 },
-          { type: 'Mascara', country: 'USA', value: 11261 }
-        ]
-      },
-      title: {
-        visible: true,
-        text: 'Unstacked area chart of cosmetic products sales'
-      },
-      stack: false,
-      xField: 'type',
-      yField: 'value',
-      zField: 'country',
-      seriesField: 'country',
-      legends: [{ visible: true, position: 'middle', orient: 'bottom' }],
-      axes: [
+      type: 'bar',
+      data: [
         {
-          orient: 'bottom',
-          mode: '3d'
-        },
-        {
-          orient: 'left',
-          mode: '3d'
-        },
-        {
-          orient: 'z',
-          mode: '3d',
-          label: { visible: true },
-          type: 'band',
-          grid: { visible: true },
-          width: 600,
-          height: 200,
-          depth: 200
+          id: 'barData',
+          values: [
+            { month: 'Monday', sales: 22 },
+            { month: 'Tuesday', sales: 13 },
+            { month: 'Wednesday', sales: 25 },
+            { month: 'Thursday', sales: 29 },
+            { month: 'Friday', sales: 38 }
+          ]
         }
       ],
-      crosshair: {
-        xField: { visible: false }
-      }
+      bar: {
+        style: {
+          texture: 'square',
+          textureSize: 6,
+          texturePadding: 1,
+          textureRatio: 0,
+          textureColor: 'orange',
+          textureOptions: {
+            dynamicTexture: (
+              ctx: any,
+              row: number,
+              column: number,
+              rowCount: number,
+              columnCount: number,
+              ratio: number,
+              graphic: any
+            ) => {
+              const _r = VRenderKits.randomOpacity(ctx, row, column, rowCount, columnCount, ratio, graphic, 0.5, 0.5);
+              ctx.globalAlpha = _r;
+              ctx.fill();
+            }
+          }
+        }
+      },
+      animationAppear: {
+        bar: {
+          channel: {
+            textureRatio: { from: 0, to: 1 }
+          },
+          easing: 'linear',
+          duration: 3000,
+          loop: true
+        }
+      },
+      xField: 'month',
+      yField: 'sales'
     };
 
     const chartSpace = new window.ChartSpace.default(spec, {
-      dom: 'container',
-      enableHtmlAttribute: true,
-      animation: false,
-      options3d: {
-        enable: true,
-        enableView3dTranform: true,
-        center: { x: 500, y: 250 }
-      }
+      dom: 'container'
     });
 
     // setTimeout(() => {